+++ packages/react-router-dom/__tests__/data-browser-router-test.tsx
@@ -1,4 +1,4 @@
-import type { ErrorResponse, Fetcher } from "@remix-run/router";
+import type { ErrorResponse, Fetcher, RouterState } from "@remix-run/router";
 import "@testing-library/jest-dom";
 import {
   act,
@@ -37,7 +37,7 @@ import {
 } from "react-router-dom";
 
 import getHtml from "../../react-router/__tests__/utils/getHtml";
-import { createDeferred } from "../../router/__tests__/utils/utils";
+import { createDeferred, tick } from "../../router/__tests__/utils/utils";
 
 testDomRouter("<DataBrowserRouter>", createBrowserRouter, (url) =>
   getWindowImpl(url, false)
@@ -7464,6 +7464,81 @@ function testDomRouter(
         fireEvent.click(screen.getByText("/d"));
         await waitFor(() => screen.getByText("D"));
         expect(spy).toHaveBeenCalledTimes(2);
+      });
+
+      it("Does not cause extra re-renders due to ViewTransitionContext updates", async () => {
+        let testWindow = getWindow("/");
+        testWindow.document.startViewTransition = (cb) => {
+          cb();
+          return {
+            ready: Promise.resolve(),
+            finished: Promise.resolve(),
+            updateCallbackDone: Promise.resolve(),
+            skipTransition: () => {},
+          };
+        };
+
+        let renders: RouterState[] = [];
+        let router = createTestRouter(
+          [
+            {
+              path: "/",
+              Component() {
+                return (
+                  <>
+                    <Link to="/page" unstable_viewTransition>
+                      /page
+                    </Link>
+                    <Outlet />
+                  </>
+                );
+              },
+              children: [
+                {
+                  index: true,
+                  async loader() {
+                    await tick();
+                    return "INDEX";
+                  },
+                  Component() {
+                    renders.push(useLocation(), useNavigation());
+                    return <h1>{useLoaderData()}</h1>;
+                  },
+                },
+                {
+                  path: "page",
+                  async loader() {
+                    await tick();
+                    return "PAGE";
+                  },
+                  Component() {
+                    renders.push(useLocation(), useNavigation());
+                    return <h1>{useLoaderData()}</h1>;
+                  },
+                },
+              ],
+            },
+          ],
+          { window: testWindow }
+        );
+        render(<RouterProvider router={router} />);
+        await waitFor(() => screen.getByText("INDEX"));
+
+        renders = [];
+        fireEvent.click(screen.getByText("/page"));
+        await waitFor(() => screen.getByText("PAGE"));
+
+        expect(renders).toMatchObject([
+          // Re-render of current location with navigation.state = "loading"
+          { pathname: "/" },
+          {
+            state: "loading",
+            location: { pathname: "/page" },
+          },
+          // Render of new location with navigation.state = "idle"
+          { pathname: "/page" },
+          { state: "idle" },
+        ]);
       });
     });
   });